pyproject.toml
[project]
name = "yaefinance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=4.0.0a1",
    "curl-cffi>=0.13.0",
    "fastapi>=0.128.4",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pydantic-settings>=2.0",
    "uvicorn>=0.40.0",
    "yfinance>=1.1.0",
]
